Update Search.js
I like the thought here, but I'm not sure we need to make this a stateful component yet. I'm not sure if the Search.js needs to remember what locations have been searched, so we'll have to research it more. 

I like having the locationArr, which may be a good substitute for using localStorage, but I don't think we need to remember every location searched. Rather, we should just remember the most recent search. This would mean that we wouldn't push new values into locationArr as much as we would redefine the location entirely. 

We can go over this more in depth later, but let me know if you have questions in the meantime!